Update Encoder.java, Decoder.java, & Main.java
Copy & pasted Lucas & Navid's code from their repo (forked from Grant; Grant's repo is forked from this repo).

Specifically, the code is a copy of the latest known commits for Encoder.java, Decoder.java, and Main.java respectively.

This is a workaround for the latest pr (merge Lucas's master branch with this master branch) because it had conflicts that weren't easy/possible to resolve.